Cleaned up scoping
Cleaned up the scoping and removed some redundant conditions in HF.40000. Did some minor general cleanup in the rest of the file. All of the references to pagan religions or religion_feature = religion_beatification in the file are unused except for HF.40025. I haven't done anything with them, but they should probably be removed or commented out.