base_hunting_events
Hunting events are no longer MTTH events, with the chances of the events occurring carefully calibrated to roughly match the originals
Moved all grand hunt events to the same file. Not sure how we want to handle the namespace.
Slightly modified the scopes of "Paranoia in the hunt" and "Courtier/vassal with hunter trait kills something" events so the chance of them firing is increased from 'basically never' to 'extremely rare'