base_married_life_events
Prevented events 30001 and 30003 from remarrying a matrillineal marriage as a patrilineal marriage.
Event 30006 now selects the same courtier for all options, and they now actually receive the bribe
The courtier also actually receives the bribe in 30008 as well.
Fixed a bug that prevented event 30030 from firing
Cleaned up inconsistencies in the 'Bored wife' event chain
Merged the Muslim version of the 'Wife offers advice: Learning' event with the regular one.