Partial rework of pagan religion head title properties
Landless-property now set on reformation, if religion head is theocratic or autocephalous. Allows these titles to be destroyed and re-created if the religion is reformed with a secular head of the religion.
'<Religion Head Title> is secular, can be destroyed and created' depends on the primary-property, but this cannot be dynamically changed, so it's disabled for all pagan religions, else both the secular and non-secular tooltip would be displayed with a non-secular head.
Added dignity to all pagan religion head titles, to match those that already had it.